Renamed module to `IfDraw2D`
