+++ src/core/brsTypes/components/RoFont.ts
@@ -5,7 +5,7 @@ import { Callable, StdlibArgument } from "../Callable";
 import { Interpreter } from "../../interpreter";
 import { Int32 } from "../Int32";
 import { FontMetrics } from "./RoFontRegistry";
-import { BrsCanvasContext2D, createNewCanvas, releaseCanvas } from "../interfaces/Draw2D";
+import { BrsCanvasContext2D, createNewCanvas, releaseCanvas } from "../interfaces/IfDraw2D";
 
 export class RoFont extends BrsComponent implements BrsValue {
     readonly kind = ValueKind.Object;
